+++ kernel/data-pipeline/dialect/mysql/src/main/java/org/apache/shardingsphere/data/pipeline/mysql/ingest/incremental/dumper/MySQLIncrementalDumper.java
@@ -200,6 +200,6 @@ public final class MySQLIncrementalDumper extends AbstractPipelineLifecycleRunna
     
     @Override
     protected void doStop() {
-        client.closeChannel();
+        client.closeChannel(true);
     }
 }